requirements.txt
numpy
requests
colorama
python-dotenv
websocket-client
blessed
urllib3
pytest
pandas
pyserial
